Move SSRF detection into detector
