+++ Aikido.Zen.Core/Sinks/DnsSink.cs
@@ -3,13 +3,14 @@ using System.Net;
 using System.Reflection;
 using System.Threading.Tasks;
 using Aikido.Zen.Core.Exceptions;
-using Aikido.Zen.Core.Helpers;
-using Aikido.Zen.Core.Models;
+using Aikido.Zen.Core.Vulnerabilities;
 
 namespace Aikido.Zen.Core.Sinks
 {
     internal static class DnsSink
     {
+        private const string OperationKind = "dns_op";
+
         [SinkFinalizer(typeof(Dns), "GetHostAddresses", "System.String")]
         [SinkFinalizer(typeof(Dns), "GetHostAddresses", "System.String", "System.Net.Sockets.AddressFamily")]
         internal static Exception OnHostAddressesResolved(string __0, IPAddress[] __result, Exception __exception, MethodBase __originalMethod)
@@ -39,41 +40,24 @@ namespace Aikido.Zen.Core.Sinks
         internal static Exception InspectResolvedAddresses(string hostname, IPAddress[] addresses, MethodBase originalMethod)
         {
             if (!OutboundRequestSink.IsRequestingOutbound() ||
-                !TryGetPrivateOrLocalIPAddress(addresses, out var privateIPAddress))
+                !SSRFDetector.TryGetPrivateOrLocalIPAddress(addresses, out var privateIPAddress))
             {
                 return null;
             }
 
-            var context = Patcher.GetContext();
-            var result = OutboundRequestSink.DetectResolvedSSRF(hostname, privateIPAddress, context);
-            if (!result.AttackKind.HasValue)
+            try
             {
+                Inspector.Inspect(
+                    originalMethod,
+                    OperationKind,
+                    context => SSRFDetector.DetectResolvedRequest(hostname, privateIPAddress, context));
+
                 return null;
             }
-
-            var operation = ReflectionHelper.GetMethodOperation(originalMethod);
-            var module = ReflectionHelper.GetMethodModule(originalMethod);
-            var blocked = !EnvironmentHelper.DryMode;
-
-            Agent.Instance.SendAttackEvent(
-                kind: result.AttackKind.Value,
-                source: result.Source,
-                payload: result.Payload,
-                operation: operation,
-                context: context,
-                module: module,
-                metadata: result.Metadata,
-                blocked: blocked,
-                paths: result.Paths);
-
-            if (context != null)
+            catch (AikidoException ex)
             {
-                context.AttackDetected = true;
+                return ex;
             }
-
-            return blocked
-                ? AikidoException.Blocked(result.AttackKind.Value, Inspector.GetBlockedOperation(operation, result))
-                : null;
         }
 
         private static async Task<IPAddress[]> InspectResolvedAddressesAsync(string hostname, Task<IPAddress[]> addressesTask, MethodBase originalMethod)
@@ -91,28 +75,6 @@ namespace Aikido.Zen.Core.Sinks
             }
 
             return addresses;
-        }
-
-        private static bool TryGetPrivateOrLocalIPAddress(IPAddress[] addresses, out string privateIPAddress)
-        {
-            privateIPAddress = null;
-
-            if (addresses == null)
-            {
-                return false;
-            }
-
-            foreach (var address in addresses)
-            {
-                var candidate = address?.ToString();
-                if (!string.IsNullOrWhiteSpace(candidate) && IPHelper.IsPrivateOrLocalIp(candidate))
-                {
-                    privateIPAddress = candidate;
-                    return true;
-                }
-            }
-
-            return false;
         }
     }
 }
