+++ Aikido.Zen.Core/Sinks/OutboundRequestSink.cs
@@ -68,63 +68,14 @@ namespace Aikido.Zen.Core.Sinks
                     });
             }
 
-            return InspectForSSRF(targetUri, context);
-        }
+            var result = SSRFDetector.DetectOutboundRequest(targetUri, context, out var inspectDns);
 
-        private static InspectionResult InspectForSSRF(Uri targetUri, Context context)
-        {
-            Uri.TryCreate(context?.Url, UriKind.Absolute, out var serverUri);
-            if (SSRFDetector.IsRequestToItself(serverUri, targetUri))
+            if (inspectDns)
             {
-                return InspectionResult.Allow();
+                EnterRequestScope();
             }
 
-            if (SSRFDetector.TryGetPrivateOrLocalIPAddress(targetUri.Host, out var privateIPAddress))
-            {
-                return DetectSSRF(targetUri, privateIPAddress, context);
-            }
-
-            EnterRequestScope();
-            return InspectionResult.Allow();
-        }
-
-        private static InspectionResult DetectSSRF(Uri targetUri, string privateIPAddress, Context context)
-        {
-            if (string.IsNullOrWhiteSpace(privateIPAddress))
-            {
-                return InspectionResult.Allow();
-            }
-
-            var hostname = targetUri.Host;
-
-            if (!SSRFDetector.IsRequestToServiceHostname(hostname) && context?.ParsedUserInput != null)
-            {
-                foreach (var userInput in context.ParsedUserInput)
-                {
-                    if (!Uri.TryCreate(userInput.Value, UriKind.Absolute, out var userUri) ||
-                        !SSRFDetector.HasSameHostAndPort(targetUri, userUri))
-                    {
-                        continue;
-                    }
-
-                    var source = UserInputHelper.GetAttackSourceFromUserInputKey(userInput.Key);
-                    return InspectionResult.Block(
-                        AttackKind.Ssrf,
-                        source: source,
-                        payload: userInput.Value,
-                        metadata: CreateMetadata(hostname, UriHelper.GetPort(targetUri), privateIPAddress),
-                        paths: new[] { UserInputHelper.GetAttackPathFromUserInputKey(userInput.Key) });
-                }
-            }
-
-            if (SSRFDetector.IsStoredSSRF(hostname, privateIPAddress))
-            {
-                return InspectionResult.Block(
-                    AttackKind.StoredSsrf,
-                    metadata: CreateMetadata(hostname, null, privateIPAddress));
-            }
-
-            return InspectionResult.Allow();
+            return result;
         }
 
         private static Uri ResolveUri(HttpRequestMessage request, HttpClient client)
@@ -142,26 +93,6 @@ namespace Aikido.Zen.Core.Sinks
             return new Uri(client.BaseAddress, request.RequestUri);
         }
 
-        private static IDictionary<string, string> CreateMetadata(string hostname, int? port, string privateIPAddress)
-        {
-            var metadata = new Dictionary<string, string>
-            {
-                { "hostname", hostname }
-            };
-
-            if (port.HasValue)
-            {
-                metadata["port"] = port.Value.ToString();
-            }
-
-            if (!string.IsNullOrWhiteSpace(privateIPAddress))
-            {
-                metadata["privateIP"] = privateIPAddress;
-            }
-
-            return metadata;
-        }
-
         private static void EnterRequestScope()
         {
             IsRequesting.Value = true;
@@ -175,56 +106,6 @@ namespace Aikido.Zen.Core.Sinks
         internal static bool IsRequestingOutbound()
         {
             return IsRequesting.Value;
-        }
-
-        internal static InspectionResult DetectResolvedSSRF(string hostname, string privateIPAddress, Context context)
-        {
-            if (string.IsNullOrWhiteSpace(privateIPAddress))
-            {
-                return InspectionResult.Allow();
-            }
-
-            if (!SSRFDetector.IsRequestToServiceHostname(hostname) && context?.ParsedUserInput != null)
-            {
-                foreach (var userInput in context.ParsedUserInput)
-                {
-                    if (!Uri.TryCreate(userInput.Value, UriKind.Absolute, out var userUri) ||
-                        !HasSameHost(hostname, userUri))
-                    {
-                        continue;
-                    }
-
-                    var source = UserInputHelper.GetAttackSourceFromUserInputKey(userInput.Key);
-                    return InspectionResult.Block(
-                        AttackKind.Ssrf,
-                        source: source,
-                        payload: userInput.Value,
-                        metadata: CreateMetadata(hostname, null, privateIPAddress),
-                        paths: new[] { UserInputHelper.GetAttackPathFromUserInputKey(userInput.Key) });
-                }
-            }
-
-            if (SSRFDetector.IsStoredSSRF(hostname, privateIPAddress))
-            {
-                return InspectionResult.Block(
-                    AttackKind.StoredSsrf,
-                    metadata: CreateMetadata(hostname, null, privateIPAddress));
-            }
-
-            return InspectionResult.Allow();
-        }
-
-        private static bool HasSameHost(string hostname, Uri uri)
-        {
-            if (!EnvironmentHelper.TrustProxy || string.IsNullOrWhiteSpace(hostname) || uri == null)
-            {
-                return false;
-            }
-
-            return string.Equals(
-                SSRFDetector.NormalizeHostname(hostname),
-                SSRFDetector.NormalizeHostname(uri.Host),
-                StringComparison.Ordinal);
         }
     }
 }
